Refactor: deposit and provisional donation entities
