+++ src/app.module.ts
@@ -42,8 +42,8 @@ import { Gift } from './entities/gift.entity';
 import { GiftogetherError } from './entities/error.entity';
 import { GiftogetherMiddleware } from './interfaces/giftogether.middleware';
 import { DepositModule } from './features/deposit/deposit.module';
-import { Deposit } from './features/deposit/domain/entities/deposit.entity';
-import { ProvisionalDonation } from './features/deposit/domain/entities/provisional-donation.entity';
+import { Deposit } from './entities/deposit.entity';
+import { ProvisionalDonation } from './entities/provisional-donation.entity';
 @Module({
   imports: [
     ScheduleModule.forRoot(),
