+++ src/features/auth/auth.module.ts
@@ -23,7 +23,7 @@ import { ImageService } from '../image/image.service';
 import { ImageInstanceManager } from '../image/image-instance-manager';
 import { TokenService } from './token.service';
 import { CreateProvisionalDonationUseCase } from '../donation/commands/create-provisional-donation.usecase';
-import { ProvisionalDonation } from '../deposit/domain/entities/provisional-donation.entity';
+import { ProvisionalDonation } from '../../entities/provisional-donation.entity';
 
 @Module({
   imports: [
