+++ src/features/deposit/commands/match-deposit.usecase.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { DepositMatchedEvent } from '../domain/events/deposit-matched.event';
 import { DepositUnmatchedEvent } from '../domain/events/deposit-unmatched.event';
-import { Deposit } from '../domain/entities/deposit.entity';
+import { Deposit } from '../../../entities/deposit.entity';
 import { GiftogetherExceptions } from '../../../filters/giftogether-exception';
-import { ProvisionalDonation } from '../domain/entities/provisional-donation.entity';
+import { ProvisionalDonation } from '../../../entities/provisional-donation.entity';
 import { DepositPartiallyMatchedEvent } from '../domain/events/deposit-partially-matched.event';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
