+++ src/features/deposit/deposit.e2e.spec.ts
@@ -4,8 +4,8 @@ import * as request from 'supertest';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { createDataSourceOptions } from 'src/tests/data-source-options';
 import { DepositModule } from './deposit.module';
-import { Deposit } from './domain/entities/deposit.entity';
-import { ProvisionalDonation } from './domain/entities/provisional-donation.entity';
+import { Deposit } from '../../entities/deposit.entity';
+import { ProvisionalDonation } from '../../entities/provisional-donation.entity';
 import { DataSource, Repository } from 'typeorm';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { GiftogetherExceptions } from 'src/filters/giftogether-exception';
