+++ src/features/deposit/deposit.module.ts
@@ -11,8 +11,8 @@ import { Funding } from 'src/entities/funding.entity';
 import { Notification } from 'src/entities/notification.entity';
 import { User } from 'src/entities/user.entity';
 import { Donation } from 'src/entities/donation.entity';
-import { Deposit } from './domain/entities/deposit.entity';
-import { ProvisionalDonation } from './domain/entities/provisional-donation.entity';
+import { Deposit } from '../../entities/deposit.entity';
+import { ProvisionalDonation } from '../../entities/provisional-donation.entity';
 import { CreateDonationUseCase } from '../donation/commands/create-donation.usecase';
 import { IncreaseFundSumUseCase } from '../funding/commands/increase-fundsum.usecase';
 import { NotificationService } from '../notification/notification.service';
