+++ src/features/deposit/domain/events/deposit-deleted.event.ts
@@ -1,4 +1,4 @@
-import { Deposit } from '../entities/deposit.entity';
+import { Deposit } from '../../../../entities/deposit.entity';
 
 export class DepositDeletedEvent {
   constructor(public readonly deposit: Deposit) {}
