+++ src/features/deposit/domain/events/deposit-event.handler.spec.ts
@@ -6,8 +6,8 @@ import { User } from 'src/entities/user.entity';
 import { Funding } from 'src/entities/funding.entity';
 import { Donation } from 'src/entities/donation.entity';
 import { Notification } from 'src/entities/notification.entity';
-import { Deposit } from '../entities/deposit.entity';
-import { ProvisionalDonation } from '../entities/provisional-donation.entity';
+import { Deposit } from '../../../../entities/deposit.entity';
+import { ProvisionalDonation } from '../../../../entities/provisional-donation.entity';
 import { ImageType } from 'src/enums/image-type.enum';
 import { AuthType } from 'src/enums/auth-type.enum';
 import { FundTheme } from 'src/enums/fund-theme.enum';
