+++ src/features/deposit/domain/events/deposit-event.handler.ts
@@ -12,7 +12,7 @@ import { IncreaseFundSumUseCase } from 'src/features/funding/commands/increase-f
 import { IncreaseFundSumCommand } from 'src/features/funding/commands/increase-fundsum.command';
 import { GiftogetherExceptions } from 'src/filters/giftogether-exception';
 import { Repository } from 'typeorm';
-import { Deposit } from '../entities/deposit.entity';
+import { Deposit } from '../../../../entities/deposit.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { FindAllAdminsUseCase } from 'src/features/admin/queries/find-all-admins.usecase';
 import { User } from 'src/entities/user.entity';
@@ -21,7 +21,7 @@ import { DepositRefundedEvent } from './deposit-refunded.event';
 import { DepositStatus } from 'src/enums/deposit-status.enum';
 import { DecreaseFundSumCommand } from 'src/features/funding/commands/decrease-fundsum.command';
 import { DepositDeletedEvent } from './deposit-deleted.event';
-import { ProvisionalDonation } from '../entities/provisional-donation.entity';
+import { ProvisionalDonation } from '../../../../entities/provisional-donation.entity';
 import { ProvisionalDonationStatus } from 'src/enums/provisional-donation-status.enum';
 
 @Injectable()
