+++ src/features/deposit/domain/events/deposit-matched.event.ts
@@ -1,5 +1,5 @@
-import { Deposit } from 'src/features/deposit/domain/entities/deposit.entity';
-import { ProvisionalDonation } from '../entities/provisional-donation.entity';
+import { Deposit } from 'src/entities/deposit.entity';
+import { ProvisionalDonation } from '../../../../entities/provisional-donation.entity';
 import { BaseEvent } from 'src/interfaces/event.interface';
 
 export class DepositMatchedEvent extends BaseEvent {
