+++ src/features/deposit/domain/events/deposit-refunded.event.ts
@@ -1,5 +1,5 @@
 import { BaseEvent } from 'src/interfaces/event.interface';
-import { Deposit } from '../entities/deposit.entity';
+import { Deposit } from '../../../../entities/deposit.entity';
 
 export class DepositRefundedEvent extends BaseEvent {
   constructor(public readonly deposit: Deposit) {
