+++ src/features/donation/commands/create-donation.command.ts
@@ -1,6 +1,6 @@
 import { User } from 'src/entities/user.entity';
 import { Funding } from 'src/entities/funding.entity';
-import { Deposit } from 'src/features/deposit/domain/entities/deposit.entity';
+import { Deposit } from 'src/entities/deposit.entity';
 
 export class CreateDonationCommand {
   constructor(
