+++ src/features/funding/funding.module.ts
@@ -22,7 +22,7 @@ import { ImageService } from '../image/image.service';
 import { S3Service } from '../image/s3.service';
 import { ImageInstanceManager } from '../image/image-instance-manager';
 import { IncreaseFundSumUseCase } from './commands/increase-fundsum.usecase';
-import { ProvisionalDonation } from '../deposit/domain/entities/provisional-donation.entity';
+import { ProvisionalDonation } from '../../entities/provisional-donation.entity';
 import { CreateProvisionalDonationUseCase } from '../donation/commands/create-provisional-donation.usecase';
 
 @Module({
