+++ src/interfaces/transition-delegate.interface.ts
